Refactor/#67 정보 동의 페이지 컴포넌트 및 custom hook 으로 상태 관리 (#87)
* refactor: consent 페이지 custom hook 구성

* refactor: consent 페이지 컴포넌트 분리 작업

* refactor: 분리했던 custom hook, component 사용
