+++ src/features/user/components/consent/index.ts
@@ -1,0 +1,2 @@
+export * from './LocationConsent';
+export * from './PrivacyConsent';
