+++ src/features/user/components/index.ts
@@ -1,3 +1,4 @@
 export * from './profile';
 export * from './profile-form';
 export * from './delete-account';
+export * from './consent';
